extracted proxy tests
